# backend/static/api.yaml
openapi: 3.0.0
info:
  title: ParkMan API
  description: A REST API for booking and managing parking lots
  version: 1.0.0
servers:
  - url: http://localhost:5000
  # description: Local Development Server
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Returns a list of all registered users. Admin access required.
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        email:
                          type: string
                          example: user@example.com
                        name:
                          type: string
                          example: Ananya Roy
                        address:
                          type: string
                          example: 123 MG Road, Bengaluru
                        pincode:
                          type: string
                          example: "560001"
                        is_admin:
                          type: boolean
                          example: false
        '404':
          description: No users found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: No user details found
        '400':
          description: User is not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Only admins can fetch all users
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch Users
                  error:
                    type: string
                    example: Database connection error
  /api/users/{id}:
    patch:
      tags:
        - Users
      summary: Edit a user's details
      description: Updates the information of a specific user. Admin access required.
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Agnibhu Mandal"
                address:
                  type: string
                  example: "123 MG Road, Kolkata"
                pincode:
                  type: string
                  example: "700001"
                isAdmin:
                  type: boolean
                  example: false
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User data edited successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User ID not found
        '400':
          description: User is not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Only admins can edit user details
        '500':
          description: Failed to update user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to edit user data
                  error:
                    type: string
                    example: SQLAlchemy Error
    put:
      tags:
        - Users
      summary: Edit user profile
      description: Updates a user's profile information including optional password update.
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - address
                - pincode
              properties:
                name:
                  type: string
                  example: "Agnibhu Mandal"
                email:
                  type: string
                  format: email
                  example: "agnibhu@example.com"
                address:
                  type: string
                  example: "123 MG Road, Kolkata"
                pincode:
                  type: string
                  example: "700001"
                password:
                  type: string
                  format: password
                  example: "new_password123"
                confirmPassword:
                  type: string
                  format: password
                  example: "new_password123"
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User profile edited successfully
        '400':
          description: Password mismatch
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Passwords don't match
        '404':
          description: User ID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User ID not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to edit user profile
                  error:
                    type: string
                    example: SQLAlchemy Exception
    delete:
      tags:
        - Users
      summary: Delete a user
      description: Deletes a user if they exist, are not an admin, and do not have active reservations.
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to delete
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User deleted successfully
        '400':
          description: Cannot delete admin or user with active reservations
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Cannot delete user who has booked reservations
        '404':
          description: User ID not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User ID not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to delete user
                  error:
                    type: string
                    example: SQLAlchemy Exception
  /api/users/admin/search:
    get:
      tags:
        - Users
      summary: Admin search for users
      description: Allows administrators to search users by name or email.
      parameters:
        - in: query
          name: userFullName
          required: false
          schema:
            type: string
          description: Full name of the user (partial match supported)
        - in: query
          name: userEmail
          required: false
          schema:
            type: string
          description: Email address of the user (partial match supported)
      responses:
        '200':
          description: Successfully found users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 7
                        email:
                          type: string
                          example: johndoe@example.com
                        name:
                          type: string
                          example: John Doe
                        address:
                          type: string
                          example: 123 MG Road, Bengaluru
                        pincode:
                          type: string
                          example: "560001"
                        is_admin:
                          type: boolean
                          example: false
        '404':
          description: No users found or no query parameters provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: User ID or name not found
        "400":
          description: User is not an admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Only admins can search this user
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Failed to fetch users
                  error:
                    type: string
                    example: Database connection error
...